JVM: move PsiSourceCompilerForInline to a separate file
